feat: Add search func
