+++ src/stores/map/useMapSearchStore.js
@@ -1,0 +1,8 @@
+import { create } from 'zustand';
+
+const useMapSearchStore = create((set) => ({
+  searchResults: [],
+  setSearchResults: (results) => set({ searchResults: results }),
+}));
+
+export default useMapSearchStore;
