+++ src/utils/DummyPlaces.js
@@ -1,0 +1,57 @@
+export const dummyPlaces = [
+  {
+    placeId: '1',
+    placeName: '강아지와 함께하는 카페',
+    category: '카페',
+    address: '서울시 강남구 테헤란로 123',
+    images: [
+      'https://via.placeholder.com/150',
+      'https://via.placeholder.com/150',
+      'https://via.placeholder.com/150',
+    ],
+    openTime: '09:00',
+    closeTime: '21:00',
+    distance: 500,
+    isLiked: true,
+  },
+  {
+    placeId: '2',
+    placeName: '강아지와 펜션에서의 휴식',
+    category: '펜션',
+    address: '경기도 가평군 청평로 456',
+    images: ['https://via.placeholder.com/150'],
+    openTime: '22:00',
+    closeTime: '10:00',
+    distance: 1200,
+    isLiked: false,
+  },
+  {
+    placeId: '3',
+    placeName: '강아지를 위한 놀이터',
+    category: '놀이터',
+    address: '부산시 해운대구 마린시티 789',
+    images: [
+      'https://via.placeholder.com/150',
+      'https://via.placeholder.com/150',
+    ],
+    openTime: '08:00',
+    closeTime: '20:00',
+    distance: 300,
+    isLiked: false,
+  },
+  {
+    placeId: '4',
+    placeName: '바닷가에서 산책',
+    category: '해수욕장',
+    address: '제주도 서귀포시 중문관광로 101',
+    images: [
+      'https://via.placeholder.com/150',
+      'https://via.placeholder.com/150',
+      'https://via.placeholder.com/150',
+    ],
+    openTime: '10:00',
+    closeTime: '22:00',
+    distance: 100,
+    isLiked: true,
+  },
+];
